Rename conditional_continuation -> max_continuous_actions
