+++ tests/test_entry_point.py
@@ -2,12 +2,12 @@ import jax.numpy as jnp
 import pytest
 from pybaum import tree_equal, tree_map
 
-from lcm.conditional_continuation import (
+from lcm.entry_point import get_lcm_function
+from lcm.input_processing import process_model
+from lcm.max_continuous_actions import (
     get_compute_conditional_continuation_policy,
     get_compute_conditional_continuation_value,
 )
-from lcm.entry_point import get_lcm_function
-from lcm.input_processing import process_model
 from lcm.state_action_space import create_state_space_info
 from lcm.utility_and_feasibility import get_utility_and_feasibility_function
 from tests.test_models import get_model_config
